scrapers: Switch to DuckDuckGo for images
Google doesn't work anymore...
Thanks to https://github.com/DarshanDeshpande/Scrapera for duckduckgo scraper
FrosT edit - Add a limit for 5 files while scraping
Will add dynamic limit later

+++ userbot/modules/scrapers.py
@@ -8,6 +8,7 @@
 import asyncio
 import json
 import os
+import glob
 import re
 import shutil
 import time
@@ -34,7 +35,7 @@ from youtube_search import YoutubeSearch
 
 from userbot import BOTLOG, BOTLOG_CHATID, CMD_HELP, TEMP_DOWNLOAD_DIRECTORY
 from userbot.events import register
-from userbot.utils import chrome, googleimagesdownload, progress
+from userbot.utils import chrome, duckduckgoscraper, progress
 
 CARBONLANG = "auto"
 TTS_LANG = "en"
@@ -106,33 +107,26 @@ async def carbon_api(e):
 @register(outgoing=True, pattern="^.img (.*)")
 async def img_sampler(event):
     """ For .img command, search and return images matching the query. """
-    await event.edit("`Processing...`")
+    await event.edit("`Processing...\n please wait for a moment...`")
     query = event.pattern_match.group(1)
-    lim = findall(r"lim=\d+", query)
-    try:
-        lim = lim[0]
-        lim = lim.replace("lim=", "")
-        query = query.replace("lim=" + lim[0], "")
-    except IndexError:
-        lim = 7
-    response = googleimagesdownload()
-
-    # creating list of arguments
-    arguments = {
-        "keywords": query,
-        "limit": lim,
-        "format": "jpg",
-        "no_directory": "no_directory"
-    }
-
-    # passing the arguments to the function
-    paths = response.download(arguments)
-    lst = paths[0][query]
-    await event.client.send_file(
-        await event.client.get_input_entity(event.chat_id), lst)
-    shutil.rmtree(os.path.dirname(os.path.abspath(lst[0])))
-    await event.delete()
-
+    scraper = duckduckgoscraper.DuckDuckGoScraper()
+    
+    #The out directory
+    os.system("mkdir -p /tmp/out/images")
+    out = ("/tmp/out/images")
+    
+    if 'query' not in locals():
+        await event.edit("Please specify a query to get images,\n like .img duck")
+    else:
+        #TODO: add a limit to the images being downloaded
+        scraper.scrape(query,1,out)
+        await asyncio.sleep(4)
+        files = glob.glob("/tmp/out/images/*.jpg")
+        await event.client.send_file(
+            await event.client.get_input_entity(event.chat_id), files
+                )
+        await event.delete()
+        os.system("rm -rf /tmp/out/images")
 
 @register(outgoing=True, pattern=r"^\.currency (.*)")
 async def moni(event):
@@ -724,8 +718,7 @@ def deEmojify(inputString):
 CMD_HELP.update(
     {
         "img": ">`.img [count] <query> [or reply]`"
-        "\nUsage: Does an image search on Google."
-        "\nCan specify the number of results needed (default is 3).",
+        "\nUsage: Does an image search on DuckDuckGo.",
         "currency": ">`.currency <amount> <from> <to>`"
         "\nUsage: Converts various currencies for you.",
         "ipinfo": ">`.ipinfo <ip_address>`"
